+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## Unreleased
 
+### Deprecated
+- Added `craft\helpers\App::parseEnv()` and `parseBooleanEnv()`, replacing their `Craft` class counterparts. ([#10319](https://github.com/craftcms/cms/discussions/10319))
+
 ### Changed
 - The content in the `resourcepaths` table is now excluded from database backups by default.
+
+### Deprecated
+- Deprecated `Craft::parseEnv()`. `craft\helpers\App::parseEnv()` should be used instead.
+- Deprecated `Craft::parseBooleanEnv()`. `craft\helpers\App::parseBooleanEnv()` should be used instead.
 
 ## 3.7.28 - 2022-01-05